moved EnumColor to ItemVampireCloak and renamed to EnumCloakColor
