+++ src/main/java/de/teamlapen/vampirism/items/ItemVampireCloak.java
@@ -1,7 +1,6 @@
 package de.teamlapen.vampirism.items;
 
 import de.teamlapen.vampirism.VampirismMod;
-import de.teamlapen.vampirism.api.EnumColor;
 import de.teamlapen.vampirism.client.model.ModelCloak;
 import de.teamlapen.vampirism.util.REFERENCE;
 
@@ -12,6 +11,7 @@ import net.minecraft.entity.EntityLivingBase;
 import net.minecraft.inventory.EntityEquipmentSlot;
 import net.minecraft.item.ItemArmor;
 import net.minecraft.item.ItemStack;
+import net.minecraft.util.IStringSerializable;
 import net.minecraft.util.NonNullList;
 import net.minecraftforge.fml.relauncher.Side;
 import net.minecraftforge.fml.relauncher.SideOnly;
@@ -42,7 +42,7 @@ public class ItemVampireCloak extends ItemArmor {
 
     public String getArmorTexture(ItemStack stack, Entity entity, EntityEquipmentSlot slot, String type) {
         return String.format(REFERENCE.MODID + ":textures/models/armor/%s/%s_%s", registeredName, registeredName,
-                EnumColor.byMetadata(stack.getMetadata()).getDyeColorName() + ".png");
+                EnumCloakColor.byMetadata(stack.getMetadata()).getDyeColorName() + ".png");
     }
 
     @Override
@@ -50,7 +50,7 @@ public class ItemVampireCloak extends ItemArmor {
         if (this.isInCreativeTab(tab))
             subItems.add(new ItemStack(this, 1, 0));
         if (tab.equals(CreativeTabs.SEARCH)) {
-            for (EnumColor s : EnumColor.values()) {
+            for (EnumCloakColor s : EnumCloakColor.values()) {
                 if (s.getMetadata() == 0)
                     continue;
                 subItems.add(new ItemStack(this, 1, s.getMetadata()));
@@ -61,6 +61,70 @@ public class ItemVampireCloak extends ItemArmor {
 
     public String getUnlocalizedName(ItemStack stack) {
         int i = stack.getMetadata();
-        return super.getUnlocalizedName() + "." + EnumColor.byMetadata(i).getUnlocalizedName();
+        return super.getUnlocalizedName() + "." + EnumCloakColor.byMetadata(i).getUnlocalizedName();
+    }
+
+    public enum EnumCloakColor implements IStringSerializable {
+        REDBLACK(0, "red_black"), BLACKRED(1, "black_red"), BLACKWHITE(2, "black_white"), WHITEBLACK(3,
+                "white_black"), BLACKBLUE(4, "black_blue");
+
+        private static final EnumCloakColor[] META_LOOKUP = new EnumCloakColor[values().length];
+        private final int meta;
+        private final String name;
+
+        private EnumCloakColor(int metaIn, String nameIn) {
+            this.meta = metaIn;
+            this.name = nameIn;
+        }
+
+        /**
+         * @return color index
+         */
+        public int getMetadata() {
+            return this.meta;
+        }
+
+        /**
+         * @return color name
+         */
+        @SideOnly(Side.CLIENT)
+        public String getDyeColorName() {
+            return this.name;
+        }
+
+        /**
+         * search for color by the given index
+         * 
+         * @param index
+         * @return color enumtype
+         */
+        public static EnumCloakColor byMetadata(int meta) {
+            if (meta < 0 || meta >= META_LOOKUP.length) {
+                meta = 0;
+            }
+
+            return META_LOOKUP[meta];
+        }
+
+        /**
+         * @return color name
+         */
+        public String getName() {
+            return this.name;
+        }
+
+        /**
+         * @return color unlocalized name
+         */
+        public String getUnlocalizedName() {
+            return this.name;
+        }
+
+        static {
+            for (EnumCloakColor enumdyecolor : values()) {
+                META_LOOKUP[enumdyecolor.getMetadata()] = enumdyecolor;
+            }
+        }
+
     }
 }
